# Optimize multimap_from_entries Presto function