Summary:
Re-use std::unordered_map between rows to avoid unnecessary memory re-allocation. Benchmark shows 20-30% speed up.

Before:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
integer_bigint##                                             1.28s   783.98m
integer_varchar##                                            1.44s   693.84m
```

After:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
integer_bigint##                                             1.00s   995.13m
integer_varchar##                                            1.17s   854.39m
```

Differential Revision: D52693295